requirements.txt
discord.py==2.3.2
lxml>=4.9.3
python-dotenv>=1.0.0